refactor: Geofence 수정사항 구현 (#30)
* refactor: Todo, Place id 자료형 Long으로 변경

* refactor: XML Empty body

* feat: Geofence 수정, 장소 추가 시 Geofence 등록 및 장소 삭제 시 Geofence 제거하도록 구현

+++ app/src/main/java/org/konkuk/placelist/place/AddTodoListener.kt
@@ -4,5 +4,5 @@ import org.konkuk.placelist.domain.Todo
 
 interface AddTodoListener {
     fun update(todo: Todo)
-    fun getTodosPlaceId(): Int
+    fun getTodosPlaceId(): Long
 }
